split core into separate files
