+++ template/assets/js/core.js
@@ -1,17 +1,16 @@
 // eslint-disable-next-line import/no-unassigned-import
 require('babel-polyfill');
 
-const url = require('url');
-const s = require('underscore.string');
-const _ = require('lodash');
 const WebFont = require('webfontloader');
 const es6promise = require('es6-promise');
-const Clipboard = require('clipboard');
 const History = require('html5-history-api');
 const jQuery = require('jquery');
 const Popper = require('popper.js');
 
+const changeHashOnScroll = require('./change-hash-on-scroll');
+const customFileInput = require('./custom-file-input');
 const ajaxForm = require('./ajax-form');
+const jumpTo = require('./jump-to');
 
 // load jQuery and Bootstrap
 // <https://stackoverflow.com/a/34340392>
@@ -56,47 +55,11 @@ WebFont.load({
   google: { families: ['Source+Code+Pro', 'Raleway', 'Montserrat'] }
 });
 
-// Remove the ugly Facebook appended hash
-// <https://github.com/jaredhanson/passport-facebook/issues/12>
-(() => {
-  if (window.location.hash !== '' && window.location.hash !== '#_=_') return;
-  window.history.replaceState(
-    undefined,
-    undefined,
-    window.location.pathname + window.location.search
-  );
-})();
+// eslint-disable-next-line import/no-unassigned-import
+require('./fixes');
 
-//
-// Since we support links containing `?return_to=/some/path`
-//
-// (e.g. if a user needs to log in to see content visible
-// that is shown further down the page, we don't want to
-// make them scroll down again, so we scroll down automatically
-// after they have successfully logged in)
-//
-// we could do ?return_to=/some/path#some-anchor
-// however this doesn't work due to server not receiving hash
-// <http://stackoverflow.com/questions/317760/how-to-get-url-hash-from-server-side>
-//
-// they can't pick-up the hash tag server-side, so we pass it
-// as a querystring parameter, therefore if there is a qs
-// parameter of `hash=` then we must convert it to location
-//
-// For more information see function `signupOrLogin` in the auth controller
-//
-// <https://nodejs.org/api/url.html#url_url_format_urlobject>
-//
-(() => {
-  const obj = url.parse(window.location.href, {
-    parseQueryString: true
-  });
-  if (!_.isObject(obj.query) || !_.isString(obj.query.hash) || s.isBlank(obj.query.hash)) return;
-  obj.hash = obj.query.hash;
-  delete obj.query.hash;
-  obj.search = undefined;
-  window.location = url.format(obj);
-})();
+// eslint-disable-next-line import/no-unassigned-import
+require('./return-to');
 
 ($ => {
   // Add pointer events polyfill
@@ -105,11 +68,11 @@ WebFont.load({
   // If the page gets resized, and on page load
   // detect the height of the top navbar and
   // set it as the `padding-top` property of body
-  function resizeNavbarPadding() {
+  const resizeNavbarPadding = () => {
     const $navbarFixedTop = $('.navbar.fixed-top');
     if ($navbarFixedTop.length === 0) return;
     $('body').css('padding-top', $navbarFixedTop.outerHeight());
-  }
+  };
 
   $(() => {
     // Resize navbar padding on load
@@ -118,7 +81,7 @@ WebFont.load({
 
     // Handle hashes when page loads
     // <http://stackoverflow.com/a/29853395>
-    if (window.location.hash !== '') jumpToTarget(window.location.hash);
+    if (window.location.hash !== '') jumpTo(window.location.hash);
 
     // Resize navbar padding when navbar is shown/collapsed
     $('.navbar-collapse').on('hidden.bs.collapse', resizeNavbarPadding);
@@ -152,36 +115,6 @@ WebFont.load({
     // Support placeholders in older browsers
     $('input, textarea').placeholder();
 
-    // Handle hashes when page scrolls
-    // <https://stackoverflow.com/a/5315993>
-    function changeHashOnScroll() {
-      // If we're at the top of the page then remove the hash
-      if (window.pageYOffset === 0)
-        return window.history.replaceState(
-          undefined,
-          undefined,
-          window.location.pathname + window.location.search
-        );
-
-      const $navbarFixedTop = $('.navbar.fixed-top');
-      const extraHeight = $navbarFixedTop.length ? $navbarFixedTop.outerHeight() : 0;
-      const $target = window.location.hash === '' ? null : $(window.location.hash);
-
-      $(':header[id]')
-        .not($target)
-        .each(function() {
-          const beginsBeforeTop = $(this).offset().top < window.pageYOffset + extraHeight;
-          const endsInVisibleArea =
-            $(this).offset().top + $(this).height() > window.pageYOffset + extraHeight;
-          if (!beginsBeforeTop || !endsInVisibleArea) return;
-          // Remove id and then add it back to prevent scroll
-          // <https://stackoverflow.com/a/1489802>
-          const id = $(this).attr('id');
-          $(this).removeAttr('id');
-          window.history.replaceState(undefined, undefined, `#${id}`);
-          $(this).attr('id', id);
-        });
-    }
     $(window).on('scroll.changeHashOnScroll', changeHashOnScroll);
 
     // Handle hashes when page loads
@@ -194,103 +127,26 @@ WebFont.load({
       $(window).load(() => {
         $hash.attr('id', hash.substring(1));
         setTimeout(() => {
-          jumpToTarget(hash);
+          jumpTo(hash);
         }, 1);
       });
     }
 
     // Handle hash change when user clicks on links
-    $('body').on('click', "a[href^='#']", function(ev) {
+    $('body').on('click', "a[href^='#']", ev => {
       ev.preventDefault();
-      jumpToTarget($(this).attr('href'));
+      jumpTo($(ev.currentTarget).attr('href'));
     });
-
-    function jumpToTarget(target) {
-      if (s.isBlank(target) || target === '#') return;
-      const $target = $(target);
-      if ($target.length === 0) return;
-      // Remove id and then add it back to prevent scroll
-      // <https://stackoverflow.com/a/1489802>
-      const id = $target.attr('id');
-      $target.removeAttr('id');
-      window.history.replaceState(undefined, undefined, `#${id}`);
-      $target.attr('id', id);
-      let offsetTop = $target.offset().top;
-      offsetTop -= Number($target.css('marginTop'));
-      offsetTop -= Number($target.css('paddingTop'));
-      if ($('.navbar.fixed-top').length > 0) offsetTop -= $('.navbar.fixed-top').outerHeight();
-      window.scrollTo(0, offsetTop);
-    }
 
     // Automatically show tooltips and popovers
     $('[data-toggle="tooltip"]').tooltip();
     $('[data-toggle="popover"]').popover();
 
     // Handle custom file inputs
-    $('body').on('change', 'input[type="file"][data-toggle="custom-file"]', function() {
-      const $input = $(this);
-      const target = $input.data('target');
-      const $target = $(target);
+    $('body').on('change', 'input[type="file"][data-toggle="custom-file"]', customFileInput);
 
-      if ($target.length === 0) return console.error('Invalid target for custom file', $input);
-
-      if (!$target.attr('data-content'))
-        return console.error('Invalid `data-content` for custom file target', $input);
-
-      // Set original content so we can revert if user deselects file
-      if (!$target.attr('data-original-content'))
-        $target.attr('data-original-content', $target.attr('data-content'));
-
-      const input = $input.get(0);
-
-      let name =
-        _.isObject(input) &&
-        _.isObject(input.files) &&
-        _.isObject(input.files[0]) &&
-        _.isString(input.files[0].name)
-          ? input.files[0].name
-          : $input.val();
-
-      if (_.isNull(name) || name === '') name = $target.attr('data-original-content');
-
-      $target.attr('data-content', name);
-    });
-
-    // Handle clipboard copy helper buttons
-    if (Clipboard.isSupported()) {
-      const clipboard = new Clipboard('[data-toggle="clipboard"]');
-      clipboard.on('success', ev => {
-        ev.clearSelection();
-        $(ev.trigger)
-          .tooltip({
-            title: 'Copied!',
-            placement: 'bottom'
-          })
-          .tooltip('show');
-        $(ev.trigger).on('hidden.bs.tooltip', () => {
-          $(ev.trigger).tooltip('dispose');
-        });
-      });
-      clipboard.on('error', ev => {
-        const key = ev.action === 'cut' ? 'X' : 'C';
-        let title = `Press <kbd>CTRL-${key}</kbd> to ${ev.action}`;
-        if (/iPhone|iPad/i.test(navigator.userAgent)) title = 'No clipboard support, sorry!';
-        else if (/Mac/i.test(navigator.userAgent))
-          title = `Press <kbd>⌘-${key}</kbd> to ${ev.action}`;
-        $(ev.trigger)
-          .tooltip({
-            title,
-            html: true,
-            placement: 'bottom'
-          })
-          .tooltip('show');
-        $(ev.trigger).on('hidden.bs.tooltip', () => {
-          $(ev.trigger).tooltip('dispose');
-        });
-      });
-    } else {
-      $('[data-toggle-clipboard]').addClass('hidden');
-    }
+    // eslint-disable-next-line import/no-unassigned-import
+    require('./clipboard');
 
     // Bind ajax form submission
     $('body').on('submit', 'form.ajax-form', ajaxForm);
