+++ template/assets/js/fixes.js
@@ -1,0 +1,10 @@
+// Remove the ugly Facebook appended hash
+// <https://github.com/jaredhanson/passport-facebook/issues/12>
+(() => {
+  if (window.location.hash !== '' && window.location.hash !== '#_=_') return;
+  window.history.replaceState(
+    undefined,
+    undefined,
+    window.location.pathname + window.location.search
+  );
+})();
